+++ fluent_pages/extensions/pagetypepool.py
@@ -7,7 +7,7 @@ from threading import Lock
 from django.contrib.contenttypes.models import ContentType
 from future.utils.six import itervalues
 from fluent_pages.models import UrlNode
-from fluent_pages.utils.load import import_apps_submodule
+from fluent_utils.load import import_apps_submodule
 from .pagetypebase import PageTypePlugin
 
 __all__ = (