+++ fluent_pages/utils/__init__.py
@@ -1,6 +1,0 @@
-"""
-The utils package only contains generic features which
-should preferably be part of Django itself.
-
-They just live in this directory to be easily bundled.
-"""